Lecture 8 - Permissions, Location, Rotation; Also refactored into packages.
