+++ app/build.gradle
@@ -35,6 +35,8 @@ dependencies {
     // From 16.1.0 -> 15.0.1
     implementation 'com.google.android.gms:play-services-maps:15.0.1'
 
+    implementation 'com.google.android.gms:play-services-location:15.0.1'
+
     implementation 'org.jetbrains.anko:anko-commons:0.10.8'
 
     // OkHttp
