+++ app/src/main/java/edu/gwu/androidtweets/ui/main/MainActivityJava.java
@@ -1,4 +1,4 @@
-package edu.gwu.androidtweets;
+package edu.gwu.androidtweets.ui.main;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -9,6 +9,9 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ProgressBar;
+
+import edu.gwu.androidtweets.R;
+import edu.gwu.androidtweets.ui.tweet.TweetsActivity;
 
 public class MainActivityJava extends AppCompatActivity {
