+++ app/src/main/java/edu/gwu/androidtweets/ui/tweet/TweetsAdapter.kt
@@ -1,4 +1,4 @@
-package edu.gwu.androidtweets
+package edu.gwu.androidtweets.ui.tweet
 
 import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
@@ -6,6 +6,8 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
+import edu.gwu.androidtweets.R
+import edu.gwu.androidtweets.data.tweet.Tweet
 
 class TweetsAdapter constructor(private val tweets: List<Tweet>) : RecyclerView.Adapter<TweetsAdapter.ViewHolder>() {
